+++ src/util/Util.cpp
@@ -169,10 +169,10 @@ std::wstring util::StrToWStr(std::string const& s) {
 
 std::string util::WStrToStr(std::wstring const& ws) {
 	std::string ret;
-	int len = WideCharToMultiByte(CP_UTF8, 0, ws.c_str(), ws.size(), NULL, 0, NULL, NULL);
+	int len = WideCharToMultiByte(CP_UTF8, 0, ws.c_str(), static_cast<int>(ws.size()), NULL, 0, NULL, NULL);
 	if (len > 0) {
 		ret.resize(len);
-		WideCharToMultiByte(CP_UTF8, 0, ws.c_str(), ws.size(), &ret[0], len, NULL, NULL);
+		WideCharToMultiByte(CP_UTF8, 0, ws.c_str(), static_cast<int>(ws.size()), &ret[0], len, NULL, NULL);
 	}
 	return ret;
 }
